% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{Add}
\alias{Add}
\title{Add}
\usage{
Add(a, b)
}
\arguments{
\item{a}{The initial value}

\item{b}{The final value}
}
\value{
The addition of the final value and the initial value
}
\description{
The function adds the inital and final value
}
