% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{angular_frequency}
\alias{angular_frequency}
\title{Angular Frequency}
\usage{
angular_frequency(f)
}
\arguments{
\item{f}{The frequency}
}
\value{
The product of pi and the frequency multoplied by 2
}
\description{
This function calculates the angular frequency
}
