% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waves&Optics.R
\name{Beat_frequency}
\alias{Beat_frequency}
\title{Beat_frequency}
\usage{
Beat_frequency(fhigh, flow)
}
\arguments{
\item{fhigh}{The high frequency}

\item{flow}{The low frequency}
}
\value{
The subtraction of fhigh from flow
}
\description{
This function calculates the beat frequency
}
