% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cav}
\alias{cav}
\title{Centripetal Acceleration with velocity}
\usage{
cav(v, r)
}
\arguments{
\item{v}{The velocity}

\item{r}{The radius}
}
\value{
The ratio of the velocity squared over the ratio
}
\description{
This function calculates the centripetal force in terms of velocity
}
