% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{caw}
\alias{caw}
\title{Centripetal Accelertion with angular velocity}
\usage{
caw(w, r)
}
\arguments{
\item{w}{The angular velocity}

\item{r}{The radius}
}
\value{
The product of the negative angular velocity squared and the radius
}
\description{
This function calculates the centripetal force in terms of angular velocity
}
