% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{delta}
\alias{delta}
\title{Delta}
\usage{
delta(xi, xf)
}
\arguments{
\item{xi}{The initial value}

\item{xf}{The final value}
}
\value{
The difference of the final value and the initial value
}
\description{
The function subtracts the 2nd value from the 1st value
}
