% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{density}
\alias{density}
\title{Density}
\usage{
density(m, v)
}
\arguments{
\item{m}{The mass of an object}

\item{v}{The volume of an object}
}
\value{
The quotient of the mass over the volume
}
\description{
This function calculates the density of an object
}
