% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{frequency}
\alias{frequency}
\title{Frequency}
\usage{
frequency(T)
}
\arguments{
\item{T}{The period}
}
\value{
The inverse of the perios
}
\description{
This function calculates the frequency by taking the inverse of the period
}
