% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waves:Optics.R
\name{Intensity}
\alias{Intensity}
\title{Intensity}
\usage{
Intensity(p, a)
}
\arguments{
\item{p}{Pressure}

\item{a}{Area}
}
\value{
The quotient of the power and area
}
\description{
This function is the Intensity with the Power divided by its area
}
