% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waves&Optics.R
\name{Intensity_level}
\alias{Intensity_level}
\title{Intensity_level}
\usage{
Intensity_level(I, I0)
}
\arguments{
\item{I}{final intensity}

\item{I0}{initial intensity}
}
\value{
The log ratio of Initial and final intensity
}
\description{
This function calculates the intentsity of a wave
}
