% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waves:Optics.R
\name{Intensity_of_Light}
\alias{Intensity_of_Light}
\title{Intensity}
\usage{
Intensity_of_Light(P, A)
}
\arguments{
\item{P}{Power}

\item{A}{Area}
}
\value{
The quotient of power over area
}
\description{
This function calculates the intensity of light
}
