% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{KE}
\alias{KE}
\title{KE}
\usage{
KE(m, v)
}
\arguments{
\item{m}{(kg) The mass of an object}

\item{v}{(m/s) The velocity at which the object is moving}
}
\value{
The product of one half of the product of mass and velocity squared
}
\description{
The function calculates the kinetic energy of an object
}
