% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thermal.R
\name{Latent_Heat}
\alias{Latent_Heat}
\title{Latent heat}
\usage{
Latent_Heat(m, L)
}
\arguments{
\item{m}{mass (kg)}

\item{L}{Latent Heat ()}

\item{Q}{Heat (kJ)}
}
\value{
The multiplication of mass and heat
}
\description{
This function calculates the Lantent Heat
}
