% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{p}
\alias{p}
\title{Momentum}
\usage{
p(m, v)
}
\arguments{
\item{m}{The mass of the object}

\item{v}{The velocity of the object}
}
\value{
The product of the mass and volume
}
\description{
The function calculates the momentum of an object
}
