% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waves:Optics.R
\name{periodic_waves}
\alias{periodic_waves}
\title{period_waves}
\usage{
periodic_waves(f, l)
}
\arguments{
\item{f}{The frquency}

\item{l}{lambda}
}
\value{
f * l
}
\description{
This function is the product of the frquency and the lambda
}
