% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pressure}
\alias{pressure}
\title{Pressure}
\usage{
pressure(f, a)
}
\arguments{
\item{f}{The force}

\item{a}{The acceleration}
}
\value{
The quotient of the force over the area
}
\description{
This function calculates the pressure of an object
}
