% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waves&Optics.R
\name{Pressure_level}
\alias{Pressure_level}
\title{Pressure_level}
\usage{
Pressure_level(P, P0)
}
\arguments{
\item{I}{final pressure}

\item{I0}{initial pressure}
}
\value{
The log ratio of Initial and final pressure
}
\description{
This function calculates the pressure of a wave
}
