% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Density}
\alias{Density}
\title{Density}
\usage{
Density(m, v)
}
\arguments{
\item{m}{(kg) The mass of an object}

\item{v}{(L^3) The volume of an object}
}
\value{
The quotient of the mass over the volume
}
\description{
This function calculates the density of an object
}
