% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Force}
\alias{Force}
\title{Force}
\usage{
Force(m, a)
}
\arguments{
\item{m}{The mass}

\item{a}{Acceleration}
}
\value{
The of the mass and acceleration
}
\description{
This function calculates the force of an object
}
