% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Frequency}
\alias{Frequency}
\title{Frequency}
\usage{
Frequency(T)
}
\arguments{
\item{T}{The period}
}
\value{
The inverse of the period
}
\description{
This function calculates the frequency by taking the inverse of the period
}
