% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ImpulseMomentum}
\alias{ImpulseMomentum}
\title{Impulse Momentum}
\usage{
ImpulseMomentum(m, v)
}
\arguments{
\item{m}{The mass}

\item{vi}{The initial velocity}

\item{vf}{The final velocity}
}
\value{
The product of mass and velocity
}
\description{
The function multiplies the mass and velocity
}
