% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waves:Optics.R
\name{Index_of_fraction}
\alias{Index_of_fraction}
\title{Index_of_fraction}
\usage{
Index_of_fraction(v)
}
\arguments{
\item{v}{phase velocity}
}
\value{
c / v
}
\description{
This function is the quotient of the speed of light divided by the phase velocity
}
