% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Power}
\alias{Power}
\title{Power}
\usage{
Power(f, v)
}
\arguments{
\item{f}{Force}

\item{v}{Velocity}
}
\value{
The product of force and velocity
}
\description{
The function multiplies the force and velocity
}
