% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Weight}
\alias{Weight}
\title{Weight}
\usage{
Weight(m)
}
\arguments{
\item{m}{The mass of the object}
}
\value{
The product of mass and the gravity
}
\description{
This function claculates the weight of an object
}
